+++ core/src/introspection/newtype.rs
@@ -7,9 +7,16 @@ use num_enum::{IntoPrimitive, TryFromPrimitive};
 use std::collections::BTreeMap;
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[cfg_attr(
+    feature = "serde",
+    derive(serde::Serialize, serde::Deserialize),
+    serde(rename_all = "kebab-case")
+)]
 pub struct Newtype {
     schema: String,
     name: String,
+
+    #[cfg_attr(feature = "serde", serde(rename = "type"))]
     target_type: TypeId,
 }